+++ CHANGELOG.md
@@ -82,3 +82,4 @@
  * Implement JWK Set Url (`jku`) in JWS headers
  * Implement Attestation JWT (`jwt`) in JWS headers
  * Implement Confirmation keys (`cnf`) in JWT
+ * Implement `CborWebToken` (RFC 8392)